Modularize folder structure
