+++ src/search/problems/index.ts
@@ -1,0 +1,2 @@
+export { SearchProblem } from './SearchProblem';
+export { SlidingTiles } from './SlidingTiles';
